requirements.txt
streamlit
streamlit_cropper
yfinance
azure-cognitiveservices-vision-computervision
msrest
Pillow
spotipy
romkan